requirements.txt
click==7.1.2
Flask==0.12.2
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
nltk==3.5
numpy==1.20.1
regex==2020.11.13
scikit-learn==0.23.1
scipy==1.6.1
textblob==0.15.3
threadpoolctl==2.1.0
tqdm==4.59.0
Werkzeug==1.0.1
gunicorn==20.0.4
pandas==1.0.5
xlrd==1.1.0
openpyxl==3.0.4
pandas-gbq==0.15.0
pandas-datareader==0.9.0
yfinance==0.1.59
google-cloud-bigquery
pyarrow
mysql-connector
statsmodels
pmdarima